refactor(EventController): thin out show() in favor of action-oriented logic
